BaseScene.h, BaseSprite::on_clicked(), BaseSprite::is_active(), global_variables.h

+++ SDL3Game/TextRect.cpp
@@ -49,7 +49,7 @@ void TextRect::change_text_color(SDL_Color new_color) {
 }
 
 
-void TextRect::change_text_size(int new_text_size) {
+void TextRect::change_text_size(float new_text_size) {
 	text_size = new_text_size;
 	SDL_DestroyTexture(texture);
 	TTF_CloseFont(font);
